+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/completion/RobotContextLevelWeigher.java
@@ -7,9 +7,10 @@ import org.jetbrains.annotations.NotNull;
 
 public class RobotContextLevelWeigher extends CompletionWeigher {
 
+    private static final int PROJECT_SCOPE_WEIGHT = 1_050;
+    private static final int LIBRARY_SCOPE_WEIGHT = 1_025;
     private static final int PARAMETER_WEIGHT = 1_000;
     private static final int VARIABLE_WEIGHT = 900;
-
     private static final int KEYWORDS_WEIGHT = 800;
 
     @Override
@@ -18,13 +19,20 @@ public class RobotContextLevelWeigher extends CompletionWeigher {
         RobotLookupContext lookupContext = element.getUserData(CompletionKeys.ROBOT_LOOKUP_CONTEXT);
         if (lookupContext == RobotLookupContext.WITHIN_KEYWORD_STATEMENT) {
             RobotLookupElementType lookupElementType = element.getUserData(CompletionKeys.ROBOT_LOOKUP_ELEMENT_TYPE);
-            if (lookupElementType == RobotLookupElementType.PARAMETER) {
+            if (lookupElementType == RobotLookupElementType.PARAMETER || lookupElementType == RobotLookupElementType.ARGUMENT) {
                 return PARAMETER_WEIGHT;
             }
             return VARIABLE_WEIGHT;
         }
         if (lookupContext == RobotLookupContext.KEYWORDS) {
             return KEYWORDS_WEIGHT;
+        }
+        if (lookupContext == RobotLookupContext.IMPORT) {
+            RobotLookupScope scope = element.getUserData(CompletionKeys.ROBOT_LOOKUP_SCOPE);
+            if (scope == RobotLookupScope.PROJECT_SCOPE) {
+                return PROJECT_SCOPE_WEIGHT;
+            }
+            return LIBRARY_SCOPE_WEIGHT;
         }
         return Integer.MIN_VALUE;
     }