+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/completion/RobotLookupContext.java
@@ -5,5 +5,5 @@ public enum RobotLookupContext {
     /**
      * Listing of keywords, usually in keyword definition or wherever you're able to place keywords
      */
-    KEYWORDS, WITHIN_KEYWORD_STATEMENT
+    KEYWORDS, WITHIN_KEYWORD_STATEMENT, IMPORT
 }