+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/completion/RobotLookupElementType.java
@@ -2,5 +2,5 @@ package dev.xeonkryptos.xeonrobotframeworkplugin.ide.completion;
 
 public enum RobotLookupElementType {
 
-    PARAMETER, KEYWORD, VARIABLE
+    PARAMETER, ARGUMENT, KEYWORD, VARIABLE
 }