+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/completion/RobotLookupScope.java
@@ -1,0 +1,6 @@
+package dev.xeonkryptos.xeonrobotframeworkplugin.ide.completion;
+
+public enum RobotLookupScope {
+
+    PROJECT_SCOPE, LIBRARY_SCOPE
+}